+++ frmts/vrt/vrtfilters.cpp
@@ -28,6 +28,9 @@
  ******************************************************************************
  *
  * $Log$
+ * Revision 1.3  2003/09/11 19:53:32  warmerda
+ * avoid type casting warnings
+ *
  * Revision 1.2  2003/08/07 17:11:21  warmerda
  * added normalized flag for kernel based filters
  *
@@ -486,12 +489,12 @@ FilterData( int nXSize, int nYSize, GDALDataType eType,
                 if( bNormalized )
                 {
                     if( dfKernSum != 0.0 )
-                        fResult = dfSum / dfKernSum;
+                        fResult = (float) (dfSum / dfKernSum);
                     else
                         fResult = 0.0;
                 }
                 else
-                    fResult = dfSum;
+                    fResult = (float) dfSum;
 
                 ((float *) pabyDstData)[iX + iY * nXSize] = fResult;
             }
